refactor: stac package
