+++ packages/topo-imagery-stac/src/topo_imagery_stac/imagery/capture_area.py
