+++ packages/topo-imagery-stac/src/topo_imagery_stac/imagery/collection.py
@@ -13,10 +13,10 @@ from topo_imagery_common.files import checksum
 from topo_imagery_common.files.files_helper import ContentType
 from topo_imagery_common.files.fs import exists, read, write
 
-from scripts.json_codec import dict_to_json_bytes
-from scripts.stac.imagery.capture_area import generate_capture_area
-from scripts.stac.imagery.collection_context import CollectionContext
-from scripts.stac.imagery.constants import (
+from topo_imagery_stac.json_codec import dict_to_json_bytes
+from topo_imagery_stac.imagery.capture_area import generate_capture_area
+from topo_imagery_stac.imagery.collection_context import CollectionContext
+from topo_imagery_stac.imagery.constants import (
     ANCILLARY_AERIAL_PHOTOS,
     ANCILLARY_NEAR_INFRARED_AERIAL_PHOTOS,
     DATA_CATEGORIES,
@@ -36,11 +36,11 @@ from scripts.stac.imagery.constants import (
     SCANNED_AERIAL_PHOTOS,
     URBAN_AERIAL_PHOTOS,
 )
-from scripts.stac.imagery.provider import Provider
-from scripts.stac.link import Link, Relation
-from scripts.stac.util.STAC_VERSION import STAC_VERSION
-from scripts.stac.util.media_type import StacMediaType
-from scripts.stac.util.stac_extensions import StacExtensions
+from topo_imagery_stac.imagery.provider import Provider
+from topo_imagery_stac.link import Link, Relation
+from topo_imagery_stac.util.STAC_VERSION import STAC_VERSION
+from topo_imagery_stac.util.media_type import StacMediaType
+from topo_imagery_stac.util.stac_extensions import StacExtensions
 
 ANY_ORTHO_AERIAL_PHOTOS = {
     ANCILLARY_AERIAL_PHOTOS,
