+++ packages/topo-imagery-stac/src/topo_imagery_stac/imagery/provider.py
