+++ packages/topo-imagery-stac/src/topo_imagery_stac/link.py
@@ -2,7 +2,7 @@ from enum import Enum
 
 from topo_imagery_common.files import checksum
 
-from scripts.stac.util.media_type import StacMediaType
+from topo_imagery_stac.util.media_type import StacMediaType
 
 
 class Relation(str, Enum):
