+++ packages/topo-imagery-stac/src/topo_imagery_stac/testing/generators.py
@@ -4,7 +4,7 @@ from typing import Callable
 
 from topo_imagery_common.files.checksum import multihash_as_hex
 
-from scripts.stac.imagery.item import STACAsset, STACProcessing, STACProcessingSoftware
+from topo_imagery_stac.imagery.item import STACAsset, STACProcessing, STACProcessingSoftware
 
 
 def fixed_now_function(now: datetime) -> Callable[[], datetime]:
