+++ packages/topo-imagery-stac/src/topo_imagery_stac/util/STAC_VERSION.py
