+++ packages/topo-imagery-stac/test/test_helpers.py
@@ -1,12 +1,8 @@
 from datetime import datetime, timedelta, timezone
-from decimal import Decimal
 from random import choice, randint
 from string import ascii_lowercase
-from typing import Iterator
 
-import pytest
 from topo_imagery_common.datetimes import format_rfc_3339_datetime_string
-from topo_imagery_stac.imagery.collection_context import CollectionContext
 
 
 def fake_linz_slug() -> str:
@@ -18,32 +14,13 @@ def fake_linz_slug() -> str:
     return f"a-random-slug-{random_string}_{start_year}-{end_year}_{gsd}m"
 
 
-@pytest.fixture
-def fake_collection_context() -> Iterator[CollectionContext]:
-    yield CollectionContext(
-        category="rural-aerial-photos",
-        domain="land",
-        region="hawkes-bay",
-        gsd=Decimal("0.3"),
-        lifecycle="completed",
-        linz_slug=fake_linz_slug(),
-        collection_id="a-random-collection-id",
-    )
-
-
 def any_epoch_datetime() -> datetime:
-    """
-    Get arbitrary datetime
-    """
     start = datetime(1970, 1, 1, tzinfo=timezone.utc)
     end = datetime(2100, 1, 1, tzinfo=timezone.utc)
     return any_datetime_between(start, end)
 
 
 def any_datetime_between(start: datetime, end: datetime) -> datetime:
-    """
-    Get arbitrary datetime between start (inclusive) and end (exclusive), with second precision.
-    """
     range_in_seconds = (end - start).total_seconds()
     return start + timedelta(seconds=randint(0, int(range_in_seconds)))
