+++ scripts/collection_from_items.py
@@ -20,11 +20,10 @@ from topo_imagery_common.files.files_helper import SUFFIX_JSON
 from topo_imagery_common.files.fs_s3 import bucket_name_from_path, get_object_parallel_multithreading, list_files_in_uri, read
 from topo_imagery_common.log.time_helper import time_in_ms
 from topo_imagery_gdal.gdal.gdal_footprint import SUFFIX_FOOTPRINT
-
-from scripts.stac.imagery.collection import CAPTURE_DATES_FILE_NAME, COLLECTION_FILE_NAME
-from scripts.stac.imagery.collection_context import CollectionContext
-from scripts.stac.imagery.constants import DATA_CATEGORIES, DATA_DOMAINS, HUMAN_READABLE_REGIONS, LAND
-from scripts.stac.imagery.create_stac import create_collection
+from topo_imagery_stac.imagery.collection import CAPTURE_DATES_FILE_NAME, COLLECTION_FILE_NAME
+from topo_imagery_stac.imagery.collection_context import CollectionContext
+from topo_imagery_stac.imagery.constants import DATA_CATEGORIES, DATA_DOMAINS, HUMAN_READABLE_REGIONS, LAND
+from topo_imagery_stac.imagery.create_stac import create_collection
 
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
