+++ scripts/standardise_validate.py
@@ -19,9 +19,8 @@ from topo_imagery_common.files.fs import exists, write
 from topo_imagery_gdal.gdal.gdal_helper import get_srs, get_vfs_path
 from topo_imagery_gdal.standardising import StandardisingConfig, run_standardising
 from topo_imagery_gdal.tiff.file_tiff import FileTiff
-
-from scripts.json_codec import dict_to_json_bytes
-from scripts.stac.imagery.create_stac import create_item
+from topo_imagery_stac.imagery.create_stac import create_item
+from topo_imagery_stac.json_codec import dict_to_json_bytes
 
 
 def get_args_parser() -> ArgumentParser:
