+++ scripts/tests/collection_from_items_test.py
@@ -11,14 +11,14 @@ from pytest import CaptureFixture, raises
 from pytest_subtests import SubTests
 from shapely.geometry import shape
 from topo_imagery_common.files.fs_s3 import write
+from topo_imagery_stac.imagery.collection import ImageryCollection
+from topo_imagery_stac.imagery.collection_context import CollectionContext
+from topo_imagery_stac.imagery.item import ImageryItem
+from topo_imagery_stac.json_codec import dict_to_json_bytes
+from topo_imagery_stac.testing.generators import any_stac_asset, any_stac_processing
 
 from scripts.collection_from_items import NoItemsError, main
 from scripts.conftest import any_epoch_datetime_string
-from scripts.json_codec import dict_to_json_bytes
-from scripts.stac.imagery.collection import ImageryCollection
-from scripts.stac.imagery.collection_context import CollectionContext
-from scripts.stac.imagery.item import ImageryItem
-from scripts.stac.imagery.tests.generators import any_stac_asset, any_stac_processing
 
 if TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
